+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## [v4.0.0](https://github.com/axlon/laravel-postal-code-validation/tree/v4.0.0)
 - Removed support for Laravel 5
+- Removed support for custom validators that do not extend the Laravel base validator
 - Removed `postal_code_for` validation rule